[@mantine/hooks] use-id: Keep id stable across effect cleanup (Activity). Fixes #8898
