+++ packages/@mantine/hooks/src/use-id/use-id.ts
@@ -1,12 +1,17 @@
-import { useId as useReactId, useState } from 'react';
+import { useId as useReactId, useRef, useState } from 'react';
 import { useIsomorphicEffect } from '../use-isomorphic-effect/use-isomorphic-effect';
 import { randomId } from '../utils';
 
 export function useId(staticId?: string) {
   const reactId = useReactId();
   const [uuid, setUuid] = useState(`mantine-${reactId.replace(/:/g, '')}`);
+  const hasInitializedRef = useRef(false);
 
   useIsomorphicEffect(() => {
+    if (hasInitializedRef.current) {
+      return;
+    }
+    hasInitializedRef.current = true;
     setUuid(randomId());
   }, []);
